Revision through web from Andreas Kupries <akupries@shaw.ca> Methods 'initalize' and 'watch' will always get full strings ('read', 'write'), and never abbreviations.
Method 'configure' broken up into 'configure', 'cget', and 'cgetall'.

Method 'initialize'. Result value changed from nested list (version +
list of methods) to plain list of methods. Versioning is through the
method names. Clarified that this implies that the signature and
semantics of 'initialize' are fixed forever, as the initial entrypoint
the C level has to know to be able to determine all the rest.

Switched from our choice of 'early binding' to use 'late binding'
instead.